requirements.txt
Django==1.9.5
django-registration==2.0.4
mccabe==0.4.0
Pillow==3.2.0
psycopg2==2.6.1
python-dateutil==2.5.2
six==1.10.0
sorl-thumbnail==12.3
dj-database-url==0.4.1
djangorestframework==3.3.3
